A06_FastAPI/requirements.txt
absl-py==2.1.0
alembic==1.13.1
aniso8601==9.0.1
annotated-types==0.6.0
anyio==4.3.0
asttokens==2.4.1
astunparse==1.6.3
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
comm==0.2.2
compress_json==1.0.10
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
docker==7.0.0
entrypoints==0.4
executing==2.0.1
fastapi==0.110.2
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.51.0
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
graphene==3.3
graphql-core==3.2.3
graphql-relay==3.2.0
greenlet==3.0.3
grpcio==1.62.2
h11==0.14.0
h5py==3.11.0
idna==3.7
importlib_metadata==7.1.0
ipykernel==6.29.4
ipython==8.23.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.0
jupyter_client==8.6.1
jupyter_core==5.7.2
keras==3.2.1
kiwisolver==1.4.5
libclang==18.1.1
Mako==1.3.3
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.3.2
mlflow==2.12.1
namex==0.0.8
nest-asyncio==1.6.0
numpy==1.26.4
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pandas==2.2.2
parso==0.8.4
pillow==10.3.0
platformdirs==4.2.0
plot_keras_history==1.1.38
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
pure-eval==0.2.2
pyarrow==15.0.2
pydantic==2.7.0
pydantic_core==2.18.1
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-multipart==0.0.9
pytz==2024.1
pywin32==306
PyYAML==6.0.1
pyzmq==26.0.2
querystring-parser==1.2.4
requests==2.31.0
rich==13.7.1
sanitize_ml_labels==1.0.51
scikit-learn==1.4.2
scipy==1.13.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.29
sqlparse==0.5.0
stack-data==0.6.3
starlette==0.37.2
support_developer==1.0.5
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
threadpoolctl==3.4.0
tornado==6.4
traitlets==5.14.3
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.29.0
waitress==3.0.0
wcwidth==0.2.13
Werkzeug==3.0.2
wrapt==1.16.0
zipp==3.18.1
